api/requirements.txt
flask_cors
requests
flasgger
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
psycopg2-binary==2.8.4
twilio
selenium>=3.141.0
Pillow
webdriver-manager
pyvirtualdisplay
python-crontab
croniter
Flask-Bcrypt==0.6.2
Flask-DebugToolbar==0.9.2
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.3.0
Flask-Script==2.0.5
Flask-Testing==0.4.2
Flask-WTF==0.11
Jinja2>=2.10.1
Mako==1.0.1
MarkupSafe==0.23
WTForms==2.0.2
Werkzeug>=0.15
alembic==0.7.4
blinker==1.3
coverage==3.7.1
itsdangerous==0.24
python-bcrypt==0.3.1
Flask-Ext